Move create_multi_provider back to rpc file
