+++ rust/services/call/host/src/test_harness.rs
@@ -4,8 +4,8 @@ use chain_client::RpcClient as RpcChainProofClient;
 use guest_wrapper::{CALL_GUEST_ELF, CHAIN_GUEST_ELF};
 use mock_chain_server::ChainProofServerMock;
 use optimism::client::factory::cached;
-use provider::{CachedMultiProvider, CachedProviderFactory};
-use rpc::rpc_urls;
+use provider::CachedMultiProvider;
+use rpc::create_multi_provider;
 pub use rpc::{block_tag_to_block_number, rpc_cache_path, rpc_cache_paths};
 pub use types::ExecutionLocation;
 
@@ -14,10 +14,6 @@ use crate::{BuilderError, Call, Config, Error, Host, PreflightResult};
 pub mod contracts;
 pub mod rpc;
 mod types;
-
-// To activate recording, set UPDATE_SNAPSHOTS to true.
-// Recording creates new test data directory and writes return data from Alchemy into files in that directory.
-const UPDATE_SNAPSHOTS: bool = false;
 
 pub async fn preflight<C>(
     test_name: &str,
@@ -112,13 +108,4 @@ fn create_host(
         op_client_factory,
         config,
     ))
-}
-
-fn create_multi_provider(test_name: &str) -> CachedMultiProvider {
-    let rpc_cache_paths = rpc_cache_paths(test_name);
-    let maybe_ethers_provider_factory =
-        UPDATE_SNAPSHOTS.then(|| provider::EthersProviderFactory::new(rpc_urls()));
-    let provider_factory =
-        CachedProviderFactory::new(rpc_cache_paths, maybe_ethers_provider_factory);
-    CachedMultiProvider::from_factory(provider_factory)
 }
